--- src/components/LeftSection/LeftSection.module.css
@@ -1,20 +1,0 @@
-.left-section {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.left-section__header {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.left-section__title-darkmode {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  align-content: flex-start;
-  gap: 1rem;
-  flex-wrap: wrap;
-}
