+++ src/components/LeftSection/Modal/ModalRegister.module.css
