+++ src/components/WorkBench/WorkBench.module.css
@@ -23,3 +23,24 @@
   overflow-y: scroll;
   min-width: 28rem;
 }
+
+.left-section {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.left-section__header {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.left-section__title-darkmode {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  align-content: flex-start;
+  gap: 1rem;
+  flex-wrap: wrap;
+}
